handel update function

+++ Controllers/bookController.js
@@ -2,9 +2,10 @@
 const bookSchema = require('./../Models/bookModel');
 
 //add new book 
-exports.addNewBook = (req, res) => {
+exports.addNewBook = (req, res,next) => {
     //create new book
     const newBook = new bookSchema({
+        _id: req.body.id,
         title: req.body.title,
         auther: req.body.auther,
         publisher: req.body.publisher,
@@ -21,47 +22,68 @@ exports.addNewBook = (req, res) => {
         res.status(201).json({
             message: 'Book Added Successfully'
         });
-    }).catch((err) => {
-        res.status(500).json({
-            message: 'Error in Adding Book'
-        });
-    });
+    }).catch((error) => next(error));
 }
 
 //update book by id
-exports.updateBookById = (req, res) => {
+exports.updateBookById = (req, res,next) => {
+    bookSchema.findById(req.params.id).then((book) => {
+        if (!book) {
+            return res.status(404).json({
+                message: 'Book Not Found'
+            });
+        }
+        bookSchema.updateOne(
+            { _id: req.params.id },
+            {
+                $set: {
+                    title: req.body.title,
+                    auther: req.body.auther,
+                    publisher: req.body.publisher,
+                    PublishingDate: req.body.PublishingDate,
+                    Category: req.body.Category,
+                    Edition: req.body.Edition,
+                    pages: req.body.pages,
+                    NoOfCopies: req.body.NoOfCopies,
+                    Avilable: req.body.Avilable,
+                    shelfNo: req.body.shelfNo
 
-    bookSchema.findByIdAndUpdate(req.params.id, {
-        title: req.body.title,
-        auther: req.body.auther,
-        publisher: req.body.publisher,
-        PublishingDate: req.body.PublishingDate,
-        Category: req.body.Category,
-        Edition: req.body.Edition,
-        pages: req.body.pages,
-        NoOfCopies: req.body.NoOfCopies,
-        Avilable: req.body.Avilable,
-        shelfNo: req.body.shelfNo
-    }).then(() => {
-        res.status(200).json({
-            message: 'Book Updated Successfully'
-        });
+                }
+            }
+        )
+       .then(() => {
+            res.status(200).json({
+                message: 'Book Updated Successfully'
+            });
+        }).catch((error) => next(error));
     }).catch((err) => {
         res.status(500).json({
-            message: 'Error in Updating Book'
+            message: 'Error in Finding Book'
         });
     });
 }
 
-//delete book by id
+
+// delete book by id
 exports.deleteBookById = (req, res) => {
-    bookSchema.findByIdAndDelete(req.params.id).then(() => {
-        res.status(200).json({
-            message: 'Book Deleted Successfully'
+    bookSchema.findById(req.params.id).then((book) => {
+        if (!book) {
+            return res.status(404).json({
+                message: 'Book Not Found'
+            });
+        }
+        bookSchema.findByIdAndDelete(req.params.id).then(() => {
+            res.status(200).json({
+                message: 'Book Deleted Successfully'
+            });
+        }).catch((err) => {
+            res.status(500).json({
+                message: 'Error in Deleting Book'
+            });
         });
     }).catch((err) => {
         res.status(500).json({
-            message: 'Error in Deleting Book'
+            message: 'Error in Finding Book'
         });
     });
 }
